Save track pause state after playing TTS
A small oversight - If the music track was paused before playing a TTS track, after TTS would end we would skip to the appropriate spot on the track and automatically play from that spot.

This change fixes that by saving the track playing state and only playing it if set before.