Made the TTS tracks have a custom title
Previously the TTS tracks didn't set their own title, which meant that LavaLink would just use the file path as the title. This resulted in the TTS Cog leaking the entire host directory structure into the temporary audio folder to any users if it had the `.audioset status` option enabled.

This change sets the track title to just `TTS` or `TTS from <user>` (if `/tts_settings say_name` is enabled)

:D