Fixed app menu exception
We had some duplicated code that wasn't changed and contained an error by making a timezone event with an invalid timezone and only checking the validity of the timezone after it was too late.

This fixes that check and makes both the commands use the same underlying function to hopefully avoid bugs like this in the future.

:D