[pastures] Configuration revamp, Slash commands, One Click Whitelisting
This revamps the configuration system to allow accessing and using multiple servers using the same bot instance. This is a breaking change and requires manual re-adding of all servers.

With this change we also fully move to slash commands as the main method of using the cog, as well as adding support for one click whitelisting.

:D